+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 4.224.5 (2022-07-06)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## 4.224.4 (2022-07-05)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## 4.224.3 (2022-07-05)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## 4.224.2 (2022-06-30)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
 ## 4.224.1 (2022-06-29)